refactor: TC-17 Move servies, guards and helpers from shared to app dir
